// frontend/common/types/enums.ts
export enum Tags {
  Biochemistry = 'Biochemistry',
  AnalyticalChemistry = 'Analytical Chemistry',
  ProteinAnalysis = 'Protein Analysis',
  ProteinQuantification = 'Protein Quantification',
  AnalyticalMethods = 'Analytical Methods',
  MolecularBiology = 'Molecular Biology',
  Virology = 'Virology',
  ProteinStructure = 'Protein Structure',
  Genetics = 'Genetics',
  DNASequencing = 'DNA Sequencing',
  EvolutionaryBiology = 'Evolutionary Biology',
  GameTheory = 'Game Theory',
  BehavioralEcology = 'Behavioral Ecology',
  Statistics = 'Statistics',
  ClinicalResearch = 'Clinical Research',
  Measurement = 'Measurement',
  CausalInference = 'Causal Inference',
  ObservationalStudies = 'Observational Studies',
  Electrophoresis = 'Electrophoresis',
  MultipleHypothesisTesting = 'Multiple Hypothesis Testing',
  DataAnalysis = 'Data Analysis',
  Pharmacology = 'Pharmacology',
  COVID19 = 'COVID-19',
  PhilosophyOfScience = 'Philosophy of Science',
  ScientificMethod = 'Scientific Method',
  ParadigmShifts = 'Paradigm Shifts',
  QuantumPhysics = 'Quantum Physics',
  TheoreticalPhysics = 'Theoretical Physics',
  Entanglement = 'Entanglement',
  Genomics = 'Genomics',
  Bioinformatics = 'Bioinformatics',
  Ecology = 'Ecology',
  ClimateScience = 'Climate Science',
  ConservationBiology = 'Conservation Biology',
  Microbiology = 'Microbiology',
  HumanHealth = 'Human Health',
  Microbiome = 'Microbiome',
  AI = 'AI',
  MedicalInformatics = 'Medical Informatics',
  DiagnosticMedicine = 'Diagnostic Medicine',
  Astrophysics = 'Astrophysics',
  Cosmology = 'Cosmology',
  ParticlePhysics = 'Particle Physics',
  Neuroscience = 'Neuroscience',
  CognitiveScience = 'Cognitive Science',
  ConsciousnessStudies = 'Consciousness Studies',
  Biotechnology = 'Biotechnology',
  GeneticEngineering = 'Genetic Engineering',
  Psychology = 'Psychology',
  SocialMediaStudies = 'Social Media Studies',
  MentalHealth = 'Mental Health',
}